aggregation items to StackView
